Move notes endpoints into new namespace
